# api/openapi.yaml
openapi: 3.0.3
info:
  title: OrbitJob API
  version: 0.1.0
  description: Code-first OpenAPI document generated from OrbitJob HTTP route metadata and DTOs.
paths:
...
  /api/v1/workflows:
    get:
      operationId: listWorkflows
      summary: List workflow definitions
      description: List the active revision of every WorkflowJob definition in the tenant. A workflow definition is a WorkflowJob custom resource, declared outside this API and read here through its projection; there is no create route. limit defaults to 50 when omitted.
      tags:
      - Workflows
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: tenant_id
        in: query
        schema:
          type: string
          minLength: 26
          maxLength: 26
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
          minimum: 1
          maximum: 100
      - name: offset
        in: query
        schema:
          type: integer
          default: 0
          minimum: 0
      responses:
        "200":
          description: Workflow list
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowListResponse"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: A group-scoped key cannot reach workflow definitions
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/workflows/{id}:
    get:
      operationId: getWorkflow
      summary: Get one workflow definition
      description: "Get the active revision of one workflow definition by revision id, with its task DAG: each task names the ScheduledJob definition it runs, the tasks that must finish first, and an optional condition over the terminal state of its dependencies."
      tags:
      - Workflows
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: tenant_id
        in: query
        schema:
          type: string
          minLength: 26
          maxLength: 26
      responses:
        "200":
          description: Workflow detail
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowGetItem"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: A group-scoped key cannot reach workflow definitions
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workflow not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/workflows/{id}/runs:
    get:
      operationId: listWorkflowRuns
      summary: List workflow runs
      description: "List one workflow's runs from the workflow ledger, newest first. Each entry carries the workflow-level phase -- the DAG's lifecycle, not any single step's -- its trigger, and the actor that asked for it."
      tags:
      - Workflows
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: tenant_id
        in: query
        schema:
          type: string
          minLength: 26
          maxLength: 26
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
          minimum: 1
          maximum: 100
      - name: offset
        in: query
        schema:
          type: integer
          default: 0
          minimum: 0
      responses:
        "200":
          description: Workflow run list
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunListResponse"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: A group-scoped key cannot reach workflow runs
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workflow not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: triggerWorkflow
      summary: Trigger a manual workflow run
      description: "Publish a WorkflowRun custom resource for a manual workflow execution. The operator materializes the workflow run row from it, row-first, and then walks the DAG: each task is an ordinary run of the ScheduledJob definition the task names. The calling credential is recorded as the run's actor, so \"who triggered it\" is answerable after the fact. Triggering a suspended workflow is a conflict. Use the X-OrbitJob-Idempotency-Key header to prevent duplicate triggers."
      tags:
      - Workflows
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: X-OrbitJob-Idempotency-Key
        in: header
        schema:
          type: string
          maxLength: 128
      responses:
        "200":
          description: Existing run returned for idempotent trigger
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTriggerResult"
        "201":
          description: Created workflow run
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTriggerResult"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: A group-scoped key cannot trigger workflows
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workflow not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Workflow is suspended
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/workflows/{id}/runs/{run_id}:
    get:
      operationId: getWorkflowRun
      summary: Get one workflow run
      description: "Get one workflow run by ledger id, with its step runs in the same response. Each step is an ordinary run of the ScheduledJob definition its task names, grouped under the workflow run; a task the walker decided not to create has no step -- its skip reason lives on the run's task decisions, visible in the phase trail."
      tags:
      - Workflows
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: run_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        "200":
          description: Workflow run detail
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunGetItem"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: A group-scoped key cannot reach workflow runs
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workflow or run not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/workflows/{id}/runs/{run_id}/cancel:
    post:
      operationId: cancelWorkflowRun
      summary: Request a workflow run cancel
      description: Request a stop for one workflow run by patching spec.cancelRequested on its WorkflowRun custom resource. The request carries no body. The operator fans the stop out over the workflow's non-terminal step runs; the workflow reaches Canceled only once its steps are observed terminal, and the response carries the phase observed at request time. Canceling a run that already finished succeeds and reports the terminal phase.
      tags:
      - Workflows
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: run_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        "200":
          description: Cancel request accepted, or the run was already terminal
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowCancelResult"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: A group-scoped key cannot cancel workflow runs
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workflow or run not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The run's WorkflowRun custom resource is missing
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    APIError:
      type: object
      properties:
        code:
          type: string
          enum:
          - MALFORMED_REQUEST
          - VALIDATION_ERROR
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - RATE_LIMITED
          - INTERNAL_ERROR
          - SERVICE_UNAVAILABLE
        field:
          type: string
        message:
          type: string
      required:
      - code
      - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/APIError"
      required:
      - error
...
    WorkflowCancelResult:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        occurrence_key:
          type: string
        phase:
          type: string
      required:
      - namespace
      - name
      - occurrence_key
      - phase
...
    WorkflowGetItem:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowTaskItem"
      required:
      - tasks
    WorkflowListItem:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
        fail_policy:
          type: string
        generation:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        name:
          type: string
        namespace:
          type: string
        schedule:
          type: string
        source_uid:
          type: string
        suspend:
          type: boolean
        task_count:
          type: integer
        timeout_seconds:
          type: integer
      required:
      - id
      - name
      - namespace
      - source_uid
      - generation
      - schedule
      - suspend
      - fail_policy
      - timeout_seconds
      - task_count
      - created_at
    WorkflowListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowListItem"
      required:
      - items
...
    WorkflowRunGetItem:
      type: object
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStepItem"
      required:
      - steps
    WorkflowRunItem:
      type: object
      properties:
        actor:
          type: string
        created_at:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        occurrence_key:
          type: string
        phase:
          type: string
        revision_id:
          type: integer
          format: int64
        trigger:
          type: string
        updated_at:
          type: string
          format: date-time
        workflow_source_uid:
          type: string
      required:
      - id
      - workflow_source_uid
      - revision_id
      - occurrence_key
      - trigger
      - actor
      - phase
      - created_at
      - updated_at
    WorkflowRunListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowRunItem"
      required:
      - items
...
    WorkflowTaskItem:
      type: object
      properties:
        condition:
          $ref: "#/components/schemas/WorkflowConditionItem"
          nullable: true
        depends_on:
          type: array
          items:
            type: string
        job_ref:
          type: string
        name:
          type: string
      required:
      - name
      - job_ref
      - depends_on
    WorkflowTriggerResult:
      type: object
      properties:
        created:
          type: boolean
        name:
          type: string
        namespace:
          type: string
        occurrence_key:
          type: string
        phase:
          type: string
        trigger:
          type: string
      required:
      - namespace
      - name
      - occurrence_key
      - trigger
      - phase
      - created